backend/requirements.txt
fastapi==0.104.1  # FastAPI核心框架（稳定版本）
uvicorn==0.24.0    # 启动FastAPI的ASGI服务器
python-multipart   # 文件上传支持
pymilvus           # Milvus 客户端
psycopg2-binary    # OpenGauss/PostgreSQL 驱动
PyPDF2             # PDF解析
pandas             # 表格处理
openpyxl           # Excel支持
sentence-transformers # 文本向量化模型
openai             # OpenAI SDK
# transformers     # CLIP模型 (可选，较大)
# torch            # PyTorch (可选，较大)
# Pillow           # 图片处理